donetask fix, not ready

+++ handlers/deletetask.go
@@ -1,7 +1,6 @@
 package handlers
 
 import (
-	"go_final_project/checkfuncs"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
@@ -11,7 +10,10 @@ import (
 func DeleteHandler(db *sqlx.DB) gin.HandlerFunc {
 	return func(c *gin.Context) {
 		id := c.Query("id")
-		checkfuncs.CheckID(id)
+		if id == "" {
+			c.JSON(http.StatusBadRequest, gin.H{"error": "Осутствует идентификатор задачи"})
+			return
+		}
 		res, err := db.Exec("DELETE FROM scheduler WHERE id = ?", id)
 		if err != nil {
 			c.AbortWithStatusJSON(http.StatusBadGateway, gin.H{"error": "Ошибка удаления данных из бд"})
@@ -23,10 +25,9 @@ func DeleteHandler(db *sqlx.DB) gin.HandlerFunc {
 			return
 		}
 		if rowsAffected == 0 {
-			c.JSON(http.StatusNotFound, gin.H{"error": "rowsaffected - 0"})
+			c.JSON(http.StatusBadRequest, gin.H{})
 			return
 		}
 		c.JSON(http.StatusOK, gin.H{})
-		return
 	}
 }
